+++ app/src/main/AndroidManifest.xml
@@ -25,16 +25,15 @@
         <activity
             android:name=".ui.component.intro.IntroActivity"
             android:exported="true">
-        </activity>
-        <activity
-            android:name=".ui.component.main.MainActivity"
-            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".ui.component.main.MainActivity"
+            android:exported="true"/>
         <activity android:name=".ui.component.profile.ProfileActivity" />
     </application>
