+++ app/src/main/java/com/msg/gcms/ui/component/clubmaker/clubDetail/MakeClubDetailFragment.kt
@@ -17,6 +17,7 @@ import coil.load
 import com.msg.gcms.R
 import com.msg.gcms.data.local.entity.ActivityPhotoType
 import com.msg.gcms.databinding.FragmentMakeClubDetailBinding
+import com.msg.gcms.ui.adapter.ActivityPhotosAdapter
 import com.msg.gcms.ui.base.BaseFragment
 import com.msg.gcms.utils.ItemDecorator
 import dagger.hilt.android.AndroidEntryPoint
